+++ src/metadata3.jl
@@ -8,10 +8,19 @@ foreach([Bool, Int8, Int16, Int32, Int64, UInt8, UInt16, UInt32, UInt64, Float16
 end
 typemap3["complex64"] = ComplexF32
 typemap3["complex128"] = ComplexF64
+typemap3["string"] = String
 
 function typestr3(t::Type)
     return lowercase(string(t))
 end
+
+function typestr3(::Type{MaxLengthString{N, UInt32}}) where {N}
+    return Dict{String, Any}(
+        "name" => "fixed_length_utf32",
+        "configuration" => Dict{String, Any}("length_bytes" => N * 4)
+        )
+end
+
 # TODO: Check raw types
 function typestr3(::Type{NTuple{N,UInt8}}) where {N}
     return "r$(N*8)"
@@ -30,8 +39,19 @@ function typestr3(s::AbstractString, codecs=nothing)
                 throw(ArgumentError("$s must describe a raw type with bit size that is a multiple of 8 bits"))
             end
         end
+
     end
     return typemap3[s]
+end
+
+typestr3(d::AbstractDict) = parse_datatype3(d)
+function parse_datatype3(d)
+    name = get(d, "name", nothing)
+
+    if name == "fixed_length_utf32"
+        return MaxLengthString{d["configuration"]["length_bytes"] ÷ 4, UInt32}
+    end
+    throw(ArgumentError("Unsupported Zarr v3 data_type: $d"))
 end
 
 function check_keys(d::AbstractDict, keys)
@@ -48,7 +68,7 @@ struct MetadataV3{T,N,P<:AbstractCodecPipeline,E<:AbstractChunkKeyEncoding} <: A
     node_type::String
     shape::Base.RefValue{NTuple{N, Int}}
     chunks::NTuple{N, Int}
-    dtype::String  # data_type in v3
+    dtype::Union{String, Dict{String, Any}}  # data_type in v3
     pipeline::P
     fill_value::Union{T, Nothing}
     chunk_key_encoding::E
@@ -69,7 +89,7 @@ Convenience constructor for MetadataV3 that builds the codec pipeline from
 and `compressor` (translated to bytes->bytes codecs).
 """
 function MetadataV3{T2,N}(zarr_format, node_type, shape::NTuple{N,Int}, chunks::NTuple{N,Int},
-        dtype::String, fill_value;
+        dtype, fill_value;
         order::Char='C',
         endian::Symbol=:little,
         compressor=BloscCompressor(),
@@ -153,6 +173,8 @@ function get_order(md::MetadataV3)
 end
 get_order(md::MetadataV2) = md.order
 
+_sizeof(x) = sizeof(x)
+_sizeof(x::Type{String}) = 1
 
 
 function Metadata3(d::AbstractDict, fill_as_missing)
@@ -209,7 +231,7 @@ function Metadata3(d::AbstractDict, fill_as_missing)
     shape = Int.(d["shape"])
 
     # Datatype
-    data_type = d["data_type"]::String
+    data_type = d["data_type"]
 
     # Chunk Grid
     chunk_grid = d["chunk_grid"]
@@ -229,7 +251,7 @@ function Metadata3(d::AbstractDict, fill_as_missing)
     T = typestr3(data_type)
     N = length(shape)
 
-    codec_ctx = (shape = shape, elsize = sizeof(Base.nonmissingtype(T)))
+    codec_ctx = (shape = shape, elsize = _sizeof(Base.nonmissingtype(T)))
     pipeline = Codecs.V3Codecs.getCodec(d["codecs"], codec_ctx)
 
     fv = fill_value_decoding(d["fill_value"], T)::T
@@ -244,7 +266,7 @@ function Metadata3(d::AbstractDict, fill_as_missing)
         node_type,
         NTuple{N, Int}(shape) |> reverse,
         NTuple{N, Int}(chunks) |> reverse,
-        data_type,
+        typestr3(T),
         pipeline,
         fv,
         chunk_key_encoding,